Fixes to help curl/test enpoints
- Changed Postgres from latest to 15 to make sure it was just using a
set version in compose
- Added a file called SecurityConfig.java - This file allows us to curl
our api when its running. Before only get would work because of a
security dependency in the project, this file disables it while we
create/test the project and we can remove it later. To curl, use the
following ' curl -v -u dev:devpass -H "Content-Type:
application/x-www-form-urlencoded" \
-d "name=Henry&age=19&email=cannonhenry019@gmail.com" \
http://localhost:8080/example 'as an example